feat(sections): refactor the mocks for the sections that requiere them
