+++ packages/authentication/src/components/SignIn/index.jsx
@@ -1,4 +1,5 @@
-import React from 'react';
+import React, { useState } from 'react';
+import { useRouter } from 'next/router';
 import {
   ButtonLogin,
   Input,
@@ -9,8 +10,35 @@ import Box from '@mui/material/Box';
 import Paper from '@mui/material/Paper';
 import { Typography } from '@mui/material';
 import Link from 'next/link';
+import users from '../../users.json';
 
 export default function SignIn() {
+  const [errorMessages, setErrorMessages] = useState({});
+  const router = useRouter();
+
+  const errors = {
+    uname: 'Usuario equivocado',
+    pass: 'Contraseña equivocada',
+  };
+
+  const handleSubmit = (event) => {
+    event.preventDefault();
+    const { uname, pass } = document.forms[0];
+    const userData = users.find((user) => user.usuario === uname.value);
+
+    if (userData) {
+      if (userData.contrasena !== pass.value) {
+        setErrorMessages({ name: 'pass', message: errors.pass });
+      } else {
+        router.push('../home');
+      }
+    } else {
+      setErrorMessages({ name: 'uname', message: errors.uname });
+    }
+  };
+
+  const errorMessage = (name) => name === errorMessages.name && errorMessages.message;
+
   return (
     <Box
       sx={{
@@ -36,20 +64,42 @@ export default function SignIn() {
           backgroundColor: 'rgb(255, 255, 255, 0.75)',
         }}
       >
-        <Typography component="h1" variant="h5" sx={{ textAlign: 'center', color: 'black' }}>Iniciar Sesión</Typography>
-        <Input label="Usuario" id="user" name="user" auto="user" size="normal" />
-        <InputPassword label="Contraseña" id="password" name="password" auto="current-password" type="password" />
-        <Box
-          display="flex"
-          justifyContent="center"
-          alignItems="center"
-          sx={{ zIndex: 1, position: 'relative', mt: 2 }}
-        >
-          <Link href="/autenticacion/contrasena" passHref>
-            <LinkButton text="¿Has olvidado tu contraseña?" />
-          </Link>
-        </Box>
-        <ButtonLogin color="secondary" type="submit" text="Entrar" href="../home" />
+        <form onSubmit={handleSubmit}>
+          <Typography
+            component="h1"
+            variant="h5"
+            sx={{ textAlign: 'center', color: 'black' }}
+          >
+            Iniciar Sesión
+          </Typography>
+          <Input
+            label="Usuario"
+            id="user"
+            name="uname"
+            auto="user"
+            size="normal"
+            errorMessage={errorMessage('uname')}
+          />
+          <InputPassword
+            label="Contraseña"
+            id="password"
+            name="pass"
+            auto="current-password"
+            type="password"
+            errorMessage={errorMessage('pass')}
+          />
+          <Box
+            display="flex"
+            justifyContent="center"
+            alignItems="center"
+            sx={{ zIndex: 1, position: 'relative', mt: 2 }}
+          >
+            <Link href="/autenticacion/contrasena" passHref>
+              <LinkButton text="¿Has olvidado tu contraseña?" />
+            </Link>
+          </Box>
+          <ButtonLogin color="secondary" type="submit" text="Entrar" />
+        </form>
       </Paper>
     </Box>
   );
