+++ packages/authentication/src/users.json
@@ -1,0 +1,82 @@
+[
+  {
+    "id": 1,
+    "usuario": "Sara",
+    "contrasena": "123456",
+    "rolId": 1,
+    "estatus": 1,
+    "actualizado": true
+  },
+  {
+    "id": 2,
+    "usuario": "Jim",
+    "contrasena": "123456",
+    "rolId": 2,
+    "estatus": 1,
+    "actualizado": true
+  },
+  {
+    "id": 3,
+    "usuario": "Pete",
+    "contrasena": "123456",
+    "rolId": 3,
+    "estatus": 1,
+    "actualizado": true
+  },
+  {
+    "id": 4,
+    "usuario": "Luis",
+    "contrasena": "123456",
+    "rolId": 4,
+    "estatus": 1,
+    "actualizado": true
+  },
+  {
+    "id": 5,
+    "usuario": "Daniel",
+    "contrasena": "123456",
+    "rolId": 5,
+    "estatus": 1,
+    "actualizado": true
+  },
+  {
+    "id": 6,
+    "usuario": "Sara",
+    "contrasena": "123456",
+    "rolId": 1,
+    "estatus": 1,
+    "actualizado": true
+  },
+  {
+    "id": 7,
+    "usuario": "Jim",
+    "contrasena": "123456",
+    "rolId": 2,
+    "estatus": 1,
+    "actualizado": true
+  },
+  {
+    "id": 8,
+    "usuario": "Pete",
+    "contrasena": "123456",
+    "rolId": 3,
+    "estatus": 1,
+    "actualizado": true
+  },
+  {
+    "id": 9,
+    "usuario": "Luis",
+    "contrasena": "123456",
+    "rolId": 4,
+    "estatus": 1,
+    "actualizado": true
+  },
+  {
+    "id": 10,
+    "usuario": "Daniel",
+    "contrasena": "123456",
+    "rolId": 5,
+    "estatus": 1,
+    "actualizado": true
+  }
+]
