--- packages/authentication/useFetch.js
@@ -1,46 +1,0 @@
-import { useEffect, useState } from 'react';
-
-function useFetch(url) {
-  const [data, setData] = useState(null);
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState(null);
-
-  useEffect(() => {
-    setLoading(true);
-    fetch
-      .get(url)
-      .then((response) => {
-        setData(response.data);
-      })
-      .catch((err) => {
-        setError(err);
-      })
-      .finally(() => {
-        setLoading(false);
-      });
-  }, [url]);
-
-  const refetch = () => {
-    setLoading(true);
-    fetch
-      .get(url)
-      .then((response) => {
-        setData(response.data);
-      })
-      .catch((err) => {
-        setError(err);
-      })
-      .finally(() => {
-        setLoading(false);
-      });
-  };
-
-  return {
-    data,
-    loading,
-    error,
-    refetch,
-  };
-}
-
-export default useFetch;
