+++ packages/solicitudes/src/components/Sections/Docentes.jsx
@@ -2,49 +2,7 @@ import { Grid } from '@mui/material';
 import { DataGrid } from '@mui/x-data-grid';
 import { ButtonAdd, Input } from '@siiges-ui/shared';
 import React from 'react';
-
-const columns = [
-  { field: 'name', headerName: 'Nombre' },
-  { field: 'type', headerName: 'Tipo' },
-  { field: 'formation', headerName: 'Formacion profecional' },
-  { field: 'asignatura', headerName: 'Asignatura' },
-  { field: 'experience', headerName: 'Experiencia' },
-  { field: 'hirability', headerName: 'Contratacion y antiguedad' },
-  { field: 'actions', headerName: 'Acciones' },
-];
-
-const rows = [
-  {
-    id: 1,
-    name: 'Jon Snow',
-    type: 'asignatura',
-    formation: 'Maestria',
-    asignatura: 'Matematicas',
-    experience: 'Resumen',
-    hirability: 'Resumen',
-    actions: 'iconos',
-  },
-  {
-    id: 2,
-    name: 'Cersei Lannister',
-    type: 'asignatura',
-    formation: 'Maestria',
-    asignatura: 'Matematicas',
-    experience: 'Resumen',
-    hirability: 'Resumen',
-    actions: 'iconos',
-  },
-  {
-    id: 3,
-    name: 'Jaime Lannister',
-    type: 'asignatura',
-    formation: 'Maestria',
-    asignatura: 'Matematicas',
-    experience: 'Resumen',
-    hirability: 'Resumen',
-    actions: 'iconos',
-  },
-];
+import { rows, columns } from './Mocks/Docentes';
 
 export default function Docentes() {
   return (
