+++ packages/solicitudes/src/components/Sections/InstitucionesAledanas.jsx
@@ -2,69 +2,7 @@ import { Grid } from '@mui/material';
 import { DataGrid } from '@mui/x-data-grid';
 import { ButtonAdd } from '@siiges-ui/shared';
 import React from 'react';
-
-const columns = [
-  { field: 'name', headerName: 'Nombre de la institucion', width: 270 },
-  { field: 'time', headerName: 'Tiempo de llegada', width: 100 },
-  { field: 'actions', headerName: 'Acciones' },
-];
-
-const rows = [
-  {
-    id: 1,
-    name: 'Jon Snow',
-    time: '9:00 a 17:00',
-    actions: 'iconos',
-  },
-  {
-    id: 2,
-    name: 'Cersei Lannister',
-    time: '9:00 a 17:00',
-    actions: 'iconos',
-  },
-  {
-    id: 3,
-    name: 'Jaime Lannister',
-    time: '9:00 a 17:00',
-    actions: 'iconos',
-  },
-  {
-    id: 4,
-    name: 'Arya Stark',
-    time: '9:00 a 17:00',
-    actions: 'iconos',
-  },
-  {
-    id: 5,
-    name: 'Daenerys Targaryen',
-    time: '9:00 a 17:00',
-    actions: 'iconos',
-  },
-  {
-    id: 6,
-    name: 'Melisandre',
-    time: '9:00 a 17:00',
-    actions: 'iconos',
-  },
-  {
-    id: 7,
-    name: 'Ferrara Clifford',
-    time: '9:00 a 17:00',
-    actions: 'iconos',
-  },
-  {
-    id: 8,
-    name: 'Rossini Frances',
-    time: '9:00 a 17:00',
-    actions: 'iconos',
-  },
-  {
-    id: 9,
-    name: 'Harvey Roxie',
-    time: '9:00 a 17:00',
-    actions: 'iconos',
-  },
-];
+import { rows, columns } from './Mocks/InstitucionesAledanas';
 
 export default function InstitucionesAledanas() {
   return (
